crash-safer: fix a needless recovery bug after success shutdown
GitHub: fix GH-386

If we (1) don't call grn_obj_flush_recursive() explicitly before we
close the DB and (2) some updates are happen in other processes after
periodical grn_obj_flush_recursive(), some .wal files are remained. It
causes unexpected DB recovery.

This commit adds the followings:

1. Ensure that other connections that use the Groonga DB are finished
   before crash-safer process closes the Groonga DB
2. Call grn_obj_flush_recursive() explicitly before crash-safer
   process closes the Groonga DB

Reported by ghevge. Thanks!!!

+++ test/helpers/sandbox.rb
@@ -46,13 +46,16 @@ module Helpers
               error_write.close
               yield(pid, input_write, output_read, error_read)
             ensure
+              finished = false
               begin
-                Process.waitpid(pid, Process::WNOHANG)
+                finished = !Process.waitpid(pid, Process::WNOHANG).nil?
               rescue SystemCallError
                 # Finished
               else
-                Process.kill(:KILL, pid)
-                Process.waitpid(pid)
+                unless finished
+                  Process.kill(:KILL, pid)
+                  Process.waitpid(pid)
+                end
               end
             end
           end
